surveillance-ui/requirements.txt
tk
grpcio
grpcio-tools
